Included the option to skip the export if the archive already exists

+++ index.php
@@ -15,7 +15,8 @@ set_time_limit(0);
 Darwin Core Archive Export</title>
 </head>
 <body style="font: 12px verdana; width: 800px;">
-<img src="images/i4life_logo_sm.jpg" width="150" height="62" style="right: 0; float: right; padding: 0 10px;" alt="i4Life">
+<img src="images/i4life_logo_sm.jpg" width="150" height="62"
+    style="right: 0; float: right; padding: 0 10px;" alt="i4Life">
 <h3>i4Life WP4 Enhanced Download Service of the Catalogue of Life:<br>
 Darwin Core Archive Export</h3>
 
@@ -32,34 +33,39 @@ if (isset($_GET['rank']) && !empty($_GET['rank']) && isset($_GET['taxon']) && !e
     // Initialize the class and check for errors
     require_once 'DCAExporter.php';
     $dcaExporter = new DCAExporter($searchCriteria);
-    $dcaExporter->useIndicator();
-    $errors = $dcaExporter->getStartUpErrors();
-    if (!empty($errors)) {
-        echo '<p><span style="color: red; font-weight: bold;">Error!</span><br>';
-        foreach ($errors as $error) {
-            echo $error . '<br>';
+    
+    // Check if archive already exits; if it does skip export
+    if (!$dcaExporter->archiveExists()) {
+        // Archive does not yet exist; create export
+        $dcaExporter->useIndicator();
+        $errors = $dcaExporter->getStartUpErrors();
+        if (!empty($errors)) {
+            echo '<p><span style="color: red; font-weight: bold;">Error!</span><br>';
+            foreach ($errors as $error) {
+                echo $error . '<br>';
+            }
+            echo "</p>\n<p><a href='index.php'>Back to the index</a></p>";
+            exit();
         }
-        echo "</p>\n<p><a href='index.php'>Back to the index</a></p>";
-        exit();
+        
+        $total = $dcaExporter->getTotalNumberOfTaxa();
+        if ($total > 0) {
+            echo "<p>Creating export for $total taxa in $rank " . ucfirst($taxon) . '.</p>';
+        }
+        else {
+            echo "<p>No results found for $rank " . ucfirst($taxon) . '. 
+                <a href="index.php">Back to the index</a></p>';
+            exit();
+        }
+        
+        echo '<p>Creating meta.xml...<br>';
+        $dcaExporter->createMetaXml();
+        echo 'Writing data to text files...<br>';
+        $dcaExporter->writeData();
+        echo '<br>Compressing to zip archive..<br>';
+        $dcaExporter->zipArchive();
+        echo "</p>\n";
     }
-    
-    $total = $dcaExporter->getTotalNumberOfTaxa();
-    if ($total > 0) {
-        echo "<p>Creating export for $total taxa in $rank " . ucfirst($taxon) . '.</p>';
-    }
-    else {
-        echo "<p>No results found for $rank " . ucfirst($taxon) . '. 
-            <a href="index.php">Back to the index</a></p>';
-        exit();
-    }
-    
-    echo '<p>Creating meta.xml...<br>';
-    $dcaExporter->createMetaXml();
-    echo 'Writing data to text files...<br>';
-    $dcaExporter->writeData();
-    echo '<br>Compressing to zip archive..<br>';
-    $dcaExporter->zipArchive();
-    echo "</p>\n";
     
     // Construct download url and calculate file size
     $ini = $dcaExporter->getExportSettings();
